fix: flask 10.17 snap breaking change fix
