+++ packages/snap/package.json
@@ -50,7 +50,7 @@
   "homepage": "https://github.com/chainsafe/filecoin-metamask-snap#readme",
   "devDependencies": {
     "@chainsafe/filsnap-types": "workspace:^",
-    "@metamask/snaps-cli": "^0.11.1",
+    "@metamask/snaps-cli": "^0.18.1",
     "@types/bn.js": "^4.11.6",
     "@types/chai": "^4.2.10",
     "@types/mocha": "^9.1.1",
