+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/Chainsafe/filsnap.git"
   },
   "source": {
-    "shasum": "yOfdreCEtwMdTmupfP+w9RyPdcOj6q1KgMfSwwd9jTg=",
+    "shasum": "uVcRkjALRav0BVsYLhRt0pWSCAyzriqpmit6Iv5TQx0=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
